requirements.txt
asgiref==3.8.1
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
anyio==4.9.0
channels==4.2.2
channels_redis==4.2.1
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
msgpack==1.1.1
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
sniffio==1.3.1
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1